backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
websockets>=11.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.21
python-multipart>=0.0.6
opencv-python>=4.8.0.76
tensorflow==2.13.0
mediapipe>=0.10.4
numpy==1.24.3
Pillow>=10.0.1
google-generativeai>=0.2.1
reportlab>=4.0.5
python-dotenv>=1.0.0
scikit-learn>=1.3.1
pandas>=2.1.1
aiofiles>=23.2.1
asyncio>=3.4.3
